+++ Classes/Service/TcaService.php
@@ -42,7 +42,12 @@ class TcaService
             'tx_yoastseo_snippetpreview' => [
                 'label' => self::LL_PREFIX_BACKEND . 'snippetPreview',
                 'exclude' => true,
-                'displayCond' => 'FIELD:tx_yoastseo_hide_snippet_preview:REQ:false',
+                'displayCond' => [
+                    'OR' => [
+                        'FIELD:tx_yoastseo_hide_snippet_preview:REQ:false',
+                        'FIELD:tx_yoastseo_disable_analysis:REQ:false',
+                    ],
+                ],
                 'config' => [
                     'type' => 'none',
                     'renderType' => 'snippetPreview',
@@ -55,7 +60,7 @@ class TcaService
             'tx_yoastseo_readability_analysis' => [
                 'label' => self::LL_PREFIX_BACKEND . 'analysis',
                 'exclude' => true,
-                'displayCond' => 'FIELD:tx_yoastseo_hide_snippet_preview:REQ:false',
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'none',
                     'renderType' => 'readabilityAnalysis',
@@ -64,7 +69,7 @@ class TcaService
             'tx_yoastseo_focuskeyword' => [
                 'label' => self::LL_PREFIX_BACKEND . 'seoFocusKeyword',
                 'exclude' => true,
-                'displayCond' => 'FIELD:tx_yoastseo_hide_snippet_preview:REQ:false',
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'input',
                 ],
@@ -72,7 +77,7 @@ class TcaService
             'tx_yoastseo_focuskeyword_analysis' => [
                 'label' => self::LL_PREFIX_BACKEND . 'analysis',
                 'exclude' => true,
-                'displayCond' => 'FIELD:tx_yoastseo_hide_snippet_preview:REQ:false',
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'none',
                     'renderType' => 'focusKeywordAnalysis',
@@ -84,6 +89,7 @@ class TcaService
             'tx_yoastseo_cornerstone' => [
                 'label' => '',
                 'exclude' => true,
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'input',
                     'default' => 0,
@@ -107,6 +113,7 @@ class TcaService
             'tx_yoastseo_focuskeyword_related' => [
                 'label' => self::LL_PREFIX_TCA . 'pages.fields.tx_yoastseo_focuskeyword_related',
                 'exclude' => true,
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'inline',
                     'foreign_table' => TableNames::RELATED_FOCUSKEYWORD,
@@ -117,6 +124,7 @@ class TcaService
             ],
             'tx_yoastseo_insights' => [
                 'exclude' => true,
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'none',
                     'renderType' => 'insights',
@@ -125,7 +133,7 @@ class TcaService
             'tx_yoastseo_focuskeyword_synonyms' => [
                 'label' => self::LL_PREFIX_TCA . 'synonyms',
                 'exclude' => true,
-                'displayCond' => 'FIELD:tx_yoastseo_hide_snippet_preview:REQ:false',
+                'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                 'config' => [
                     'type' => 'input',
                 ],
@@ -159,6 +167,14 @@ class TcaService
             ],
         ];
         if ($this->table === TableNames::PAGES) {
+            $columns['tx_yoastseo_disable_analysis'] = [
+                'label' => self::LL_PREFIX_BACKEND . 'disableAnalysis',
+                'exclude' => true,
+                'onChange' => 'reload',
+                'config' => [
+                    'type' => 'check',
+                ],
+            ];
             $columns['tx_yoastseo_hide_snippet_preview'] = [
                 'label' => self::LL_PREFIX_BACKEND . 'hideSnippetPreview',
                 'exclude' => true,
@@ -181,6 +197,7 @@ class TcaService
                 'tx_yoastseo_facebook_preview' => [
                     'label' => self::LL_PREFIX_TCA . 'pages.fields.tx_yoastseo_facebook_preview',
                     'exclude' => true,
+                    'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                     'config' => [
                         'type' => 'none',
                         'renderType' => 'facebookPreview',
@@ -189,6 +206,7 @@ class TcaService
                 'tx_yoastseo_twitter_preview' => [
                     'label' => self::LL_PREFIX_TCA . 'pages.fields.tx_yoastseo_twitter_preview',
                     'exclude' => true,
+                    'displayCond' => 'FIELD:tx_yoastseo_disable_analysis:REQ:false',
                     'config' => [
                         'type' => 'none',
                         'renderType' => 'twitterPreview',
@@ -263,8 +281,10 @@ class TcaService
         ExtensionManagementUtility::addFieldsToPalette(
             $this->table,
             'yoast-advanced',
-            '--linebreak--, tx_yoastseo_hide_snippet_preview'
+            '--linebreak--, tx_yoastseo_disable_analysis, tx_yoastseo_hide_snippet_preview'
         );
+        $GLOBALS['TCA'][$this->table]['palettes']['yoast-advanced']['description']
+            = self::LL_PREFIX_TCA . 'pages.palettes.advanced.description';
     }
 
     protected function addToTypes(): void
@@ -295,7 +315,7 @@ class TcaService
 
         ExtensionManagementUtility::addToAllTCAtypes(
             $this->table,
-            '--palette--;' . self::LL_PREFIX_TCA . 'pages.palettes.advances;yoast-advanced,',
+            '--palette--;' . self::LL_PREFIX_TCA . 'pages.palettes.advanced;yoast-advanced,',
             $this->types,
             'after: sitemap_priority'
         );