+++ Resources/Public/JavaScript/yoast-plugin.js
@@ -19,8 +19,10 @@ class YoastPlugin {
         this.initializeState();
         new FocusKeyphrase().init();
         new SnippetPreview().init();
-        new StatusIcon().init();
-        new AnalysisResult().init();
+        if (YoastConfiguration.isAnalysisEnabled()) {
+            new StatusIcon().init();
+            new AnalysisResult().init();
+        }
         DocumentService.ready().then(() => {
             this.previewRequest(configuration.requestData);
         });
@@ -46,6 +48,9 @@ class YoastPlugin {
                 return;
             }
             store.setState({ content: data });
+            if (!YoastConfiguration.isAnalysisEnabled()) {
+                return;
+            }
             await analysis.refresh();
             this.saveScores();
             this.saveProminentWords();