+++ Tests/Unit/Service/Ajax/ProminentWordsHandlerTest.php
@@ -11,8 +11,13 @@ declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\Tests\Unit\Service\Ajax;
 
+use Doctrine\DBAL\Result;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Database\Query\Expression\ExpressionBuilder;
+use TYPO3\CMS\Core\Database\Query\QueryBuilder;
+use TYPO3\CMS\Core\Database\Query\Restriction\QueryRestrictionContainerInterface;
 use TYPO3\CMS\Core\Http\JsonResponse;
 use TYPO3\CMS\Core\Http\ServerRequest;
 use TYPO3\CMS\Core\Http\Stream;
@@ -31,7 +36,34 @@ class ProminentWordsHandlerTest extends UnitTestCase
         parent::setUp();
 
         $this->prominentWordsService = $this->createMock(ProminentWordsService::class);
-        $this->subject = new ProminentWordsHandler($this->prominentWordsService);
+        $connectionPool = $this->createConnectionPoolMock(['tx_yoastseo_disable_analysis' => 0]);
+        $this->subject = new ProminentWordsHandler($this->prominentWordsService, $connectionPool);
+    }
+
+    protected function createConnectionPoolMock(array|false $fetchResult): ConnectionPool
+    {
+        $result = $this->createMock(Result::class);
+        $result->method('fetchAssociative')->willReturn($fetchResult);
+
+        $expressionBuilder = $this->createMock(ExpressionBuilder::class);
+        $expressionBuilder->method('eq')->willReturn('');
+
+        $restrictions = $this->createMock(QueryRestrictionContainerInterface::class);
+        $restrictions->method('removeAll')->willReturnSelf();
+        $restrictions->method('add')->willReturnSelf();
+
+        $queryBuilder = $this->createMock(QueryBuilder::class);
+        $queryBuilder->method('select')->willReturnSelf();
+        $queryBuilder->method('from')->willReturnSelf();
+        $queryBuilder->method('where')->willReturnSelf();
+        $queryBuilder->method('expr')->willReturn($expressionBuilder);
+        $queryBuilder->method('getRestrictions')->willReturn($restrictions);
+        $queryBuilder->method('executeQuery')->willReturn($result);
+
+        $connectionPool = $this->createMock(ConnectionPool::class);
+        $connectionPool->method('getQueryBuilderForTable')->willReturn($queryBuilder);
+
+        return $connectionPool;
     }
 
     protected function createRequestWithJsonBody(array $data): ServerRequest