+++ ext_tables.sql
@@ -6,6 +6,7 @@ CREATE TABLE pages (
     tx_yoastseo_focuskeyword_synonyms tinytext,
     tx_yoastseo_focuskeyword_related int(11) DEFAULT '0' NOT NULL,
     tx_yoastseo_hide_snippet_preview tinyint(3) DEFAULT '0' NOT NULL,
+    tx_yoastseo_disable_analysis tinyint(3) DEFAULT '0' NOT NULL,
     tx_yoastseo_cornerstone tinyint(3) DEFAULT '0' NOT NULL,
     tx_yoastseo_score_readability varchar(50) DEFAULT '' NOT NULL,
     tx_yoastseo_score_seo varchar(50) DEFAULT '' NOT NULL,